+++ tests/test_stripe.py
@@ -8,16 +8,14 @@
 import os
 
 import pytest
-from django.utils.timezone import now
 from shuup.front.basket import get_basket
-from shuup.testing.factories import (
-    create_order_with_product, get_default_product, get_default_shop,
-    get_default_supplier, get_default_tax_class)
+from shuup.testing.factories import get_default_shop, get_default_tax_class
 from shuup.utils.excs import Problem
-from shuup.utils.http import retry_request
 
 from shuup_stripe.checkout_phase import StripeCheckoutPhase
 from shuup_stripe.models import StripeCheckoutPaymentProcessor
+
+from .utils import create_order_for_stripe, get_stripe_token
 
 
 @pytest.fixture
@@ -32,47 +30,13 @@ def stripe_payment_processor():
         secret_key=sk, publishable_key="x")
 
 
-def _create_order_for_stripe(stripe_payment_processor):
-    product = get_default_product()
-    supplier = get_default_supplier()
-    order = create_order_with_product(
-        product=product, supplier=supplier, quantity=1,
-        taxless_base_unit_price=100, tax_rate=0
-    )
-    payment_method = stripe_payment_processor.create_service(
-        None, shop=order.shop, tax_class=get_default_tax_class(), enabled=True)
-    order.payment_method = payment_method
-    order.cache_prices()
-    assert order.taxless_total_price.value > 0
-    if not order.payment_data:
-        order.payment_data = {}
-    order.save()
-    return order
-
-
-def get_stripe_token(stripe_payment_processor):
-    return retry_request(
-        method="post",
-        url="https://api.stripe.com/v1/tokens",
-        auth=(stripe_payment_processor.secret_key, "x"),
-        data={
-            "card[number]": "4242424242424242",
-            "card[exp_month]": 12,
-            "card[exp_year]": now().date().year + 1,
-            "card[cvc]": 666,
-        }
-    ).json()
-
-
-##############################################################################
-
 @pytest.mark.django_db
 def test_stripe_basics(rf, stripe_payment_processor):
     """
     :type rf: RequestFactory
     :type stripe_payment_module: StripeCheckoutModule
     """
-    order = _create_order_for_stripe(stripe_payment_processor)
+    order = create_order_for_stripe(stripe_payment_processor)
     token = get_stripe_token(stripe_payment_processor)
     token_id = token["id"]
     order.payment_data["stripe"] = {"token": token_id}
@@ -86,7 +50,7 @@ def test_stripe_basics(rf, stripe_payment_processor):
 
 @pytest.mark.django_db
 def test_stripe_bogus_data_fails(rf, stripe_payment_processor):
-    order = _create_order_for_stripe(stripe_payment_processor)
+    order = create_order_for_stripe(stripe_payment_processor)
     order.payment_data["stripe"] = {"token": "ugubugudugu"}
     order.save()
     with pytest.raises(Problem):
@@ -127,13 +91,13 @@ def test_stripe_checkout_phase(rf, stripe_payment_processor):
 
 @pytest.mark.django_db
 def test_stripe_checkout_phase_with_misconfigured_module(rf):
-    stripe_payment_prossor = (
+    stripe_payment_processor = (
         StripeCheckoutPaymentProcessor.objects.create())
     request = rf.get("/")
     request.shop = get_default_shop()
     request.session = {}
     request.basket = get_basket(request)
-    service = stripe_payment_prossor.create_service(
+    service = stripe_payment_processor.create_service(
         None, shop=request.shop, tax_class=get_default_tax_class(),
         enabled=True)
     checkout_phase = StripeCheckoutPhase(
